services/ai-processor/requirements.txt
httpx>=0.28.0
redis>=5.2.0
psycopg2-binary>=2.9.0
elasticsearch>=8.17.0
python-dotenv>=1.0.0
pydantic>=2.10.0
lightrag>=1.0.0
numpy>=2.2.0
